Adding lib.rs due https://github.com/rust-lang/cargo/issues/15231
